+++ bukkit/src/main/java/de/maxhenkel/voicechat/compatibility/Compatibility1_19.java
@@ -4,7 +4,9 @@ import com.mojang.brigadier.arguments.ArgumentType;
 import de.maxhenkel.voicechat.BukkitVersion;
 import org.bukkit.entity.Player;
 
-public class Compatibility1_19 extends BaseCompatibility {
+import static de.maxhenkel.voicechat.compatibility.ReflectionUtils.*;
+
+public class Compatibility1_19 extends JsonMessageBaseCompatibility {
 
     public static final BukkitVersion VERSION_1_19 = BukkitVersion.parseBukkitVersion("1.19-R0.1");
 
@@ -48,13 +50,13 @@ public class Compatibility1_19 extends BaseCompatibility {
     private void send(Player player, String json, boolean status) {
         Object entityPlayer = callMethod(player, "getHandle");
         Object playerConnection = getField(entityPlayer, "b");
-        Class<?> packet = getClass("net.minecraft.network.protocol.Packet");
+        Class<?> packet = getClazz("net.minecraft.network.protocol.Packet");
         Class<?> craftChatMessage = getBukkitClass("util.CraftChatMessage");
 
-        Class<?> iChatBaseComponentClass = getClass("net.minecraft.network.chat.IChatBaseComponent");
+        Class<?> iChatBaseComponentClass = getClazz("net.minecraft.network.chat.IChatBaseComponent");
         Object iChatBaseComponent = callMethod(craftChatMessage, "fromJSON", new Class[]{String.class}, json);
 
-        Class<?> clientboundSystemChatPacketClass = getClass("net.minecraft.network.protocol.game.ClientboundSystemChatPacket");
+        Class<?> clientboundSystemChatPacketClass = getClazz("net.minecraft.network.protocol.game.ClientboundSystemChatPacket");
 
         Object clientboundSystemChatPacket = callConstructor(clientboundSystemChatPacketClass, new Class[]{iChatBaseComponentClass, int.class}, iChatBaseComponent, getId(player, status));
 
@@ -63,17 +65,17 @@ public class Compatibility1_19 extends BaseCompatibility {
 
     private int getId(Player player, boolean status) {
         Object world = callMethod(player.getWorld(), "getHandle");
-        Class<?> iRegistryCustomClass = getClass("net.minecraft.core.IRegistryCustom");
+        Class<?> iRegistryCustomClass = getClazz("net.minecraft.core.IRegistryCustom");
         Object iRegistryCustom = callMethod(world, "s");
 
-        Class<?> resourceKeyClass = getClass("net.minecraft.resources.ResourceKey");
+        Class<?> resourceKeyClass = getClazz("net.minecraft.resources.ResourceKey");
 
-        Class<?> iRegistry = getClass("net.minecraft.core.IRegistry");
+        Class<?> iRegistry = getClazz("net.minecraft.core.IRegistry");
         Object resourceKey = getField(iRegistry, "bI");
 
         Object registry = callMethod(iRegistryCustomClass, iRegistryCustom, "d", new Class[]{resourceKeyClass}, resourceKey);
 
-        Class<?> chatMessageTypeClass = getClass("net.minecraft.network.chat.ChatMessageType");
+        Class<?> chatMessageTypeClass = getClazz("net.minecraft.network.chat.ChatMessageType");
         Object chatMessageType;
         if (status) {
             chatMessageType = getField(chatMessageTypeClass, "d");