+++ bukkit/src/main/java/de/maxhenkel/voicechat/compatibility/Compatibility1_20_2.java
@@ -4,7 +4,9 @@ import com.mojang.brigadier.arguments.ArgumentType;
 import de.maxhenkel.voicechat.BukkitVersion;
 import org.bukkit.entity.Player;
 
-public class Compatibility1_20_2 extends BaseCompatibility {
+import static de.maxhenkel.voicechat.compatibility.ReflectionUtils.*;
+
+public class Compatibility1_20_2 extends JsonMessageBaseCompatibility {
 
     public static final BukkitVersion VERSION_1_20_2 = BukkitVersion.parseBukkitVersion("1.20.2-R0.1");
 
@@ -37,19 +39,19 @@ public class Compatibility1_20_2 extends BaseCompatibility {
 
     @Override
     public ArgumentType<?> playerArgument() {
-        Class<?> argumentEntity = getClass("net.minecraft.commands.arguments.ArgumentEntity");
+        Class<?> argumentEntity = getClazz("net.minecraft.commands.arguments.ArgumentEntity");
         return callMethod(argumentEntity, "c");
     }
 
     @Override
     public ArgumentType<?> uuidArgument() {
-        Class<?> argumentEntity = getClass("net.minecraft.commands.arguments.ArgumentUUID");
+        Class<?> argumentEntity = getClazz("net.minecraft.commands.arguments.ArgumentUUID");
         return callMethod(argumentEntity, "a");
     }
 
     private void send(Player player, String json, boolean status) {
         Object entityPlayer = callMethod(player, "getHandle");
-        Class<?> iChatBaseComponentClass = getClass("net.minecraft.network.chat.IChatBaseComponent");
+        Class<?> iChatBaseComponentClass = getClazz("net.minecraft.network.chat.IChatBaseComponent");
         Class<?> craftChatMessage = getBukkitClass("util.CraftChatMessage");
         Object iChatBaseComponent = callMethod(craftChatMessage, "fromJSON", new Class[]{String.class}, json);
         callMethod(entityPlayer, "a", new Class[]{iChatBaseComponentClass, boolean.class}, iChatBaseComponent, status);