+++ bukkit/src/main/java/de/maxhenkel/voicechat/integration/commodore/CommodoreCommands.java
@@ -11,12 +11,18 @@ import me.lucko.commodore.Commodore;
 public class CommodoreCommands {
 
     public static void registerCompletions(Commodore commodore) {
+        ArgumentType<?> playerArgument = playerArgument();
+        ArgumentType<?> uuidArgument = uuidArgument();
+        if (playerArgument == null || uuidArgument == null) {
+            //TODO
+            return;
+        }
         LiteralArgumentBuilder<?> literalBuilder = LiteralArgumentBuilder.literal(VoiceChatCommands.VOICECHAT_COMMAND)
                 .then(LiteralArgumentBuilder.literal("help"))
-                .then(LiteralArgumentBuilder.literal("test").then(RequiredArgumentBuilder.argument("target", playerArgument())))
-                .then(LiteralArgumentBuilder.literal("invite").then(RequiredArgumentBuilder.argument("target", playerArgument())))
-                .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group_id", uuidArgument())))
-                .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group_id", uuidArgument()).then(RequiredArgumentBuilder.argument("password", StringArgumentType.string()))))
+                .then(LiteralArgumentBuilder.literal("test").then(RequiredArgumentBuilder.argument("target", playerArgument)))
+                .then(LiteralArgumentBuilder.literal("invite").then(RequiredArgumentBuilder.argument("target", playerArgument)))
+                .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group_id", uuidArgument)))
+                .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group_id", uuidArgument).then(RequiredArgumentBuilder.argument("password", StringArgumentType.string()))))
                 .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group_name", StringArgumentType.string()).suggests(GroupNameSuggestionProvider.INSTANCE)))
                 .then(LiteralArgumentBuilder.literal("join").then(RequiredArgumentBuilder.argument("group_name", StringArgumentType.string()).suggests(GroupNameSuggestionProvider.INSTANCE).then(RequiredArgumentBuilder.argument("password", StringArgumentType.string()))))
                 .then(LiteralArgumentBuilder.literal("leave"));